Added crasher to crash meteor app.
At this point I've tried several things, simple and complicated to build
a system for hashing and salting with different variables, and
environment keys, and obfuscators. It was becoming needlessly too
complicated, so I took a step back and decided to not write any code for
this at this point.

All the things I have yet tried for obfuscating server side code hasn't
worked. It always crash meteor app. I have tried to pass strings, have
different package.js files, different obfuscators but none worked.

Best we can do to protect premium code of nucleus is to make it as
inaccessible as possible for the user. We are not providing keys to the
user, so they can never log in directly to the instance. Only interface
they have to access the filesystem is through the terminal we provide in
nucleus. We can configure the terminal to make every command that tries
to access the directory where premium_nucleus is symlinked from
fail. This we will do when we change nucleuside.com for launching
instances with users' apps and install nucleus on them.